Event Handling: Moved handleChange to TodosLogic. Passed down to TodosList and TodoItem
