+++ src/components/TodoItem.jsx
@@ -1,18 +1,4 @@
-const TodoItem = ({ itemProp }) => {
-    const handleChange = (id) => {
-        setTodos((prevState) =>
-            prevState.map((todo) => {
-                if (todo.id === id) {
-                    return {
-                        ...todo,
-                        completed: !todo.completed,
-                    };
-                }
-                return todo;
-            })
-        );
-    };
-
+const TodoItem = ({ itemProp, handleChange }) => {
     return (
         <li>
             <input
