+++ src/components/TodosList.jsx
@@ -1,14 +1,14 @@
-import React from 'react';
-import TodoItem from '@/components/TodoItem';
-
-const TodosList = ({ todosProps, setTodos }) => {
+const TodosList = ({ todosProps, handleChange }) => {
     return (
         <ul>
             {todosProps.map((todo) => (
-                <TodoItem key={todo.id} itemProp={todo} setTodos={setTodos} />
+                <TodoItem
+                    key={todo.id}
+                    itemProp={todo}
+                    handleChange={handleChange}
+                />
             ))}
         </ul>
     );
 };
-
 export default TodosList;
