+++ src/components/TodosLogic.jsx
@@ -5,27 +5,25 @@ import TodosList from '@/components/TodosList';
 
 const TodosLogic = () => {
     const [todos, setTodos] = useState([
-        {
-            id: 1,
-            title: 'Setup development environment',
-            completed: true,
-        },
-        {
-            id: 2,
-            title: 'Develop website and add content',
-            completed: false,
-        },
-        {
-            id: 3,
-            title: 'Deploy to live server',
-            completed: false,
-        },
+        // ...
     ]);
-
+    const handleChange = (id) => {
+        setTodos((prevState) =>
+            prevState.map((todo) => {
+                if (todo.id === id) {
+                    return {
+                        ...todo,
+                        completed: !todo.completed,
+                    };
+                }
+                return todo;
+            })
+        );
+    };
     return (
         <div>
             <InputTodo />
-            <TodosList todosProps={todos} setTodos={setTodos} />
+            <TodosList todosProps={todos} handleChange={handleChange} /> {/* Pass handleChange here */}
         </div>
     );
 };
